chore(wallet-sdk-example) cleanup
